modificazione della struttura del progetto per essere piu modulare
